backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.5.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
bcrypt==4.1.2
google-generativeai==0.3.2
